I was unable to write out the wanted_list.csv file
The wanted_list.csv file is now written out.  It outputs the correct information (as far as I could tell).  However, I recommend testing on Linux.  I modified it on Windows, and changed the way the file is opened.  There's still an issue - if there are commas in the fields that will generate another column upon import (as expected).  I would have to look a little more at the standard method of dealing with commas on CSVs to make a recommendation here.